Tailorized "2005-02-04 18:02:25 by logg" Add (very) simple divergence test to homotopy test example.
An old problem has reappeared: PETSc reports

 ** On entry to DGEMV  parameter number  6 had an illegal value

when the homotopy test problem (src/demo/solver/ode/homotopy/simple) tries
to solve for the second starting point.

Does anyone know why? The problem disappears if the program is run
through valgrind so it seems likely that it is a memory allocation error,
but valgrind reports no errors...