Tailorized "2005-10-18 17:11:27 by logg" Automatically reinitialize GMRES solver when system size changes.
When creating a GMRES object and repeatedly calling solve(), the PETSc
KSP pointer will be automatically reinitialized when the system size changes.
Looks like PETSc initialize some temporary data only on the first solve and
doesn't check if the system size changes on a subsequent solve.

This seems to do the trick for the multi-adaptive Newton solver (but more
tests are needed).