Tailorized "2005-10-23 05:38:05 by logg" Add directory and template for test programs.
This file is to used for testing out new features implemented
in the library, which means that the contents of this file
might be (should be) constantly changing. Anything can be
thrown into this file at any time and may also be removed
at any time. Use this for temporary test programs that are
not suitable to be implemented as demos in src/demo.