Tailorized "2005-11-01 20:59:09 by johanjan" Heat:
Added new module "heat" as a demo and testbed for the PDE<->ODE solver
interface. This also includes a Python script "visualize_heat.py"
(requires Mayavi 1.4) for visualizing the solution on a mesh and a
Python script "plot_heat.py" for plotting the solution as components
over time (ODE-style).

Convdiff:

Added a Python script "visualize_convdiff.py" which does the same
thing as above. Reduced the time step and total time to better be able
to visualize the solution.