Tailorized "2005-11-20 15:19:12 by johanjan" ODE:
Fixed bug in mcG(q). Previously, f at the right-hand endpoint was
saved even if the slab was rejected. This is now fixed.

The error for mcG(1) in the benchmark has been reduced by up to one
magnitude (runtime is unchanged). The error is still larger than for
cG(1) though, so there appears to remain at least one source of
error. We'll have to keep looking a bit more.