Tailorized "2005-11-29 19:56:02 by logg" Improve and restructure implementation of the Function class
The Function class has been broken to a hierarchy of classes. The Function
class now functions as an envelope class and all functionality is implemented
in the respective letter class (sub classes of GenericFunction) for each
type of function. This should make it easier to add new functionality in the
future. The logic of the Function class is also more transparent.

All previous functionality is retained, but note the following minor
changes to the interface:

1. User-defined functions should overload

       real operator() (const Point& p, unsigned int i)

both for scalar and vector-valued functions. (Just ignore the component
index i for a scalar function.)

2. User-defined functions can now also be defined directly from a pointer
to a function of the type real(*FunctionPointer)(const Point& p, unsigned int).

3. User-defined functions can be saved to file if a mesh is attached to
the function before saving:

    u.attach(mesh);

Note that the vector dimension needs to be specified in the constructor
of a user-defined function for vector-valued functions.

4. User-defined functions no longer have access to _cell. Instead,
user-defined functions that depend on the current cell can call the function
cell(), which returns the current cell during assembly and generates an
error otherwise (when there is no current cell).

5. The problem that Andy noted earlier (non-matching elements for repeated
assembly with the same function and different elements) should now be
fixed.

Next, I plan to add support for picking sub functions of mixed functions and
components of vector-valued functions.

Please try it out and tell me if I broke anything.