Tailorized "2005-12-20 20:06:33 by logg" New parameter name for solution file: "solution file name".
Default value is "solution.pvd". All modules should obey this
parameter if possible or declare new parameters for the solution file
name(s) so that the output format and file name can be specified from
a user program.

The following now works for the Poisson solver:

    PoissonSolver poisson(mesh, f, bc);
    poisson.set("solution file name", "solution.m");
    poisson.solve();

Specifying a parameter with poisson.set() only affects the parameter
in the database local to the PoissonSolver object. If the parameter
is not specified, the value is inherited from the global database.

Temporarily changed the name of the parameter for output file name for
the ODE solvers to "ode solution file name". The default for the ODE
solvers needs to be different from "solution.pvd" (should be "solution.py"
or "solution.m"). This will be fixed when the ODE solvers have moved
to localized parameter databases: ode.set("...", ...).