Tailorized "2006-02-01 15:45:45 by johanjan" Build:
Introduced libtool into the build system. In addition to the old
libraries (one for every kernel component and module), a single DOLFIN
static and shared library is built. This removes the need for the
repetition of libraries when linking.

Note that when linking against shared libraries on Linux, the path of
the shared libraries must be known to the dynamic loader
(ld.so). These paths are specified in /etc/ld.so.conf or with the
environment variable LD_LIBRARY_PATH.