Add post-proceessing step to compute and output strain in /src/demo/pde/elasticty.
The computation of the strain is very slow because in the constructor of PDE, the linear solver set to direct, and it's not possible for the user to chanage this before the solve is peformed.  

The VTK format can handle tensors, so when Function supports tensors, the processing will become simpler.