Implement new strategy for conservative modification of time steps.
This is a poor man's dual weighting function. For each component,
we look at all other components that the component in question
depends on and tell these other components to reduce their time
steps to the same level. A similar effect would be accomplished
by weighting with the proper weight from the dual solution, but
until we (re-)implement the solution of the dual, this seems to
be a good solution.

Also changed the partitioning threshold from 0.2 to 0.5 for benchmark.

This further speeds up the multi-adaptive benchmark and restores the
the proportionality e ~ tol.