Implement ILU preconditioner for uBlas Krylov solvers.
The uBlas-based GMRES and BiCGStab solvers with ILU preconditioning run about 20% faster than the corresponding PETSc version. The preconditioner interface needs to be unified with the PETSc preconditioner interface.

Something is up with the assembly of uBlas sparse matrices for the demo programs. It's very slow. Need to look at this because the assembly in src/test/main.cpp is very fast.