Add support for setting periodic boundary conditions.
To set a periodic boundary condition, specify a pair of domains
G and H by a SubDomain. The inside() function should specify the
points of G and the map() function should specify a mapping F : H --> G.

Then create a PeriodicBoundaryCondition and apply() it to the linear system.

Bugs/features:

  1. Only works for scalar subsystems, so it needs to be done once for
  each subsystem of a vector function.

  2. Setting the entries in the matrix can perhaps be optimized.

  3. PeriodicBoundaryCondition is not a subclass of BoundaryCondition,
  so it does not work to give a PeriodicBoundaryCondition as argument
  to LinearPDE.

To fix 3, should we rename BoundaryCondition --> DirichletBC
and have PeriodicBC and DirichletBC as subclasses of BoundaryCondition?