Run MTL4 sandbox through dolfin::Assembler interface.
Results are still very good, espcially for Navier-Stokes. MTL4 relies on specifying the number of non-zeroes per row. Without this, it is slower than uBLAS.

For Navier-Stokes (matrix size 681k x 681k), MTL4 with number of non zeroes specified is 3.2 times faster than uBLAS, and 2 times faster for the re-assmeble.

Comparison is not entirely fair because eventually some form of sparsity pattern will need to be computed for MTL4 to provide number of non-zeroes per row, although knowing the sparsity pattern may possible speed insertion.

MTL4 is 1.25 faster for the matrix-vector multiplication test.