Work on getting Function and FunctionSpace in place: 1. Remove all member data for Assembler 2. Make all functions in Assembler static 3. Update in assembler.h 4. Remove assembler.cpp 5. Start work on using const Mesh& everywhere.
The use of const Mesh& is not complete and heavy editing is needed
everywhere to get this in place. The problem is that the iterators
have so far assumed that the Mesh is non-const. From now on, all
iterators will iterate over a const Mesh and the Mesh may consequently
not be modified through the iterators.

assembler.h is just a copy of the functions in Assembler.h so
we could possibly use some trick to put the functions in Assembler::
in the dolfin:: namespace