Python callback functions now works.
We are not using simple_array, but rather a specific director typemap that
depends on that it only kicks in for Function::eval(double* values, const double *x).

Therefore I had %ignore the SubDomain::inside(const double* x, bool on_boundary)
functions which is not used in the python interface.