Heavy work on the python interface. Base line: we can now assemble a tensor.
Extended the compile function capabilities, and added thorough (I think ;))
documentation in Function docstring.

Assemble.py is up to date with the new FunctionSpace interface. A tensor is now
assembled using:

  A = assemble(a)

An interface for precompiled ufc::forms is provided but not tested, a la suggestion
from Martin previously.

Poisson demo is a bit messy but it shows some of the new functionality.