Fix issue with the help text which appeared when multiple windows was opened.
For some reason we need to make sure the help text actor is not destroyed be the event loop is entered.
Also note that VTK is very picky about the order of the destruction of objects. Wrong order leads to a seg fault. 
(I'm not sure if this is bad design on our side, bad design in VTK or a bug in VTK).