Rename vertex_coordinates -> coordinate_dofs many places.
In particular, this renaming should cover everywhere the meaning
is to get the degrees of freedom for the coordinates on a cell,
but not the places where the actual vertex coordinates are used.
This is preparation for non-affine meshes.

Add get_coordinate_dofs besides get_vertex_coordinates to Cell.
Remove pointer version, everyone uses the std::vector version.