Use LocalAssembler in LocalSolver
Using LocalAssembler adds support for facet integrals in LocalSolver.
The included unit test fails due to a bug in LocalAssembler which is
fixed in the next commit.

See pull request #244 for benchmark comparisons. The local matrices
are changed from row major to column major (Eigen default), but there
does not seem to be a performance impact of this in the preliminary
benchmark at least.