Add test to see that we fixed issue 696
We create a facet function and a ds Measure using this facet function.
We need not mark any facets and only use marker 0 (which is the
default). This was enough to make the previous implementation assemble
the form incorrectly. With the fixed LocalSolver RHS domains are handled
correctly and the correct local tensor is assembled leading to the test
not failing.