Ensure that SNES solver works when invoked multple times
This is because PETSc does not allow to set preconditioning Jacobian
back from A to P which is empty and fail starangely in
SNESSetFromOptions. Test in test_PETScSNES_solver.py.

FIXME: This fix is semantically incorrect because when new
preconditioner is given with new nonlinear problem argument,
preconditioner is not reset. But it's strange use case.