Improve parallel scaling in PointIntegralSolver
This commit implements a PETSc performance optimization for the
PointIntegralSolver to achieve better parallel scaling. More specifically,
we use the VEC_IGNORE_OFF_PROC_ENTRIES flag to avoid cross-processor
communication during VecAssembly* calls.

This flag may be used here, because the PointIntegralSolver.step method
only set local values of the to-be-assembled PETSc vector. Note: this
flag does not affect the update_ghost calls, so the ghost values will
still be valid.

I have verified the correctness of the new implementation by comparing
the outputs of the state solution of a non-trivial PDE-ODE test case
(cardiac example of the PDE-ODE paper by ME Rognes et al) for both with
and without this flag in parallel (16 cores) and obtained the same
results down to solver tolerance (here 1e-10).