Update documentation (mostly adding links)
The documentation is now a bit better than a lot of empty pages. There
is probably no need for duplication of information, so I have added
links to our web pages

Also some cleanup of the C++ API docs. The Python docs are still
missing and the demos give a ton of Sphinx warnings so they should also
be fixed.

Also: documenting_dolfin_api.rst should be updated!