This finally compiles, now it just segfaults :-p
Tested without mpi4py support (not yet detected in CMake?). The code
then compiles, but segfaults during the unit tests.

There are a lot of wraper updates in this commit and the previous with
some likelyhood of copy/paste errors like returning the wrong class from
a constructor ... should be checked when the branch starts working.

All the new constructors currently returns unique_ptr. This may not be
needed? It is what is used in the pybind11 docs and I did not want more
problems than the necessary minimum. Any API endpoint exposing MPI_Com
without wrapping allready leads to a wall of compiler error messages.

If mpi4py is available the custom caster should automatically cast to
the mpi4py Python object (from the custom C++ comm wrapper object). If
mpi4py is not present then yet another wrapper is returned, this one
with a pybind11 class_<>... definition which homefully will work like
an opaque object in Python, working in the dolfin API, but not
interoperable with petsc4py/mpi4py