DOLFINX wrapper for ufc_expression (#1083)
This allows DOLFINX to use FFCX generated code to evaluate any UFL Expression at a fixed set of points on the reference cell. This could be used to calculate e.g. derivatives of a Function at quadrature points, that can subsequently be passed to a non-FEniCS routine to calculate a constitutive update.