Fix #1558.
Issue with previous attempt at fixing this code was that https://github.com/FEniCS/dolfinx/blob/main/cpp/dolfinx/geometry/utils.cpp#L279
was not adapted to the initial entity being the actual leaf index, and not the node of the bbox tree.
This would cause suprious nans in computation of initial distances and segfault.