Let dlx.la.Vector manage the scope of the petsc4py wrapper (#1) (#3092)
* Let dlx.la.Vector manage the scope of the petsc4py wrapper (#1)

This PR allows for a `dlx.la.Vector` (rather than a `dlx.fem.Function`) to manage the memory and the scope of the `petsc4py.PETSc.vec` object that wraps the underlying data.

The implementation of the property `dlx.fem.Function.vector` was also modified to delegate the creation of the wrapper to the underlying `dlx.la.Vector`.

All changes are backward compatible.

* Update python/dolfinx/fem/function.py

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>

* Update python/dolfinx/fem/function.py

Remove # type: ignore

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>

* Deprecate dlx.fem.Function.vector; replace it with dlx.fem.Function.x.petsc_vec

* Fix ruff checks

* Fix ruff checks import order

* Ruff formatt fix

* Update function.py

Fix spelling in warnings message.

* Update function.py

Specify that it is a deprecation warning

* Ruff formatting error

---------

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>